refactor(lib): simplify app folder structure
