+++ src/lib/app/utils/homePageData.svelte.js
