+++ src/lib/app/utils/searchBarIcoglyphs.svelte.js
