+++ src/routes/icon/[icoGlyphNameSlug]/mainIcoG.svelte
@@ -1,6 +1,6 @@
 <script>
 	import icoGlyphs from '$lib/index.js';
-	import appState from '$lib/app/core/stores/appState.svelte.js';
+	import appState from '$lib/app/appState.svelte.js';
 
 	let { data } = $props();
